Updates loops in client.js
The for loop can use `let` for the variable names. Caching the length up
front saves work over the loop.

For...in loops use an arbitrary order, so shouldn't be used for looping
where order matters. For...of loops also only iterate over the members
of the iterable so the `hasOwnProperty` check is not required.